refactor: move unpauseMentoToken out of utils
